Move WorkspaceMenu to common because it is used by both the container and the list view.
